Implemented Home page header
I spent waaaaay too much time trying to figure out how to customize the MUI components just to give in and use the sx styling instead.
Oh well. I ended up learning how to make a custom theme with the MUI library which was pretty cool in my opinion.
Anyways, future my will be making the logo function as a home button and also the about/sign in buttons. However, the main focus for today will just be getting all the elements on one page :'(