Initial skeletons of Log in/Sign up forms
I want to allow users to sign up/login with authentication but also be able to customize my own design for these forms so I won't be focusing too much on using Auth0. I might use it if I decide I don't care for the design because it would be a lot easier with these prebuilt authentication forms. However, I'm going to figure out a way around it. I'm currently trying to instantiate the backend so I can pass the user info to my database.