Updated sign up
I didn't get much time today but I checked that the sign up and login code works through Postman. I need to look more into the CI Integration test (I think that's what it was called on the Postman documentation) but for now I just want to get all the basics done. Now I'm currently working on getting the UI to match the wireframes I made. I will continue to work on it throughout the week