Updated email signup page
The page now has the main logo which doubles as a home button in the top right corner, password covering/uncovering, and a better sizing for the image on the left. This page still struggles with the sizing dynamically but that will be fixed as I continue to work on the page.