Added spotify authentication
FINALLYgit add . I was able to add code that allows users to sign in with Spotify. it was a little confusing to figure out at first, but then I realized I was using the callback url instead of the login url on my onClick function for my button (haha silly me).
Next up is integrating google mapsgit add . (and building practically the whole home page)